requirements.txt
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
matplotlib>=3.8.0
jupyter>=1.0.0
tabulate>=0.9.0
ipython>=8.0.0
